+++ x/cronos/keeper/ibc_hooks_test.go
@@ -75,7 +75,6 @@ func (suite *KeeperTestSuite) TestOnRecvVouchers() {
 				suite.app.GetSubspace(types.ModuleName),
 				suite.app.BankKeeper,
 				keepertest.IbcKeeperMock{},
-				suite.app.GravityKeeper,
 				suite.app.EvmKeeper,
 			)
 			suite.app.CronosKeeper = cronosKeeper